separated the navigator, added another state into the store, disable animation for navigation, perf and responses improve alot
