+++ app/containers/FootSectionRouterContainer.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import FootSectionRouter from './../components/navigators/footSection/FootSectionRouter.js';
 
 const mapStateToProps = state => ({
-  selectedTab: state.selectedTab
+  nav: state.nav
 })
 
 export default connect(mapStateToProps)(FootSectionRouter)
