+++ app/store/store.js
@@ -1,21 +1,32 @@
-import { createStore } from 'redux'
+import { createStore, combineReducers } from 'redux'
+import FootSectionNav from './../components/navigators/footSection/FootSectionNav.js'
 
-const initState = {
-  selectedTab: 'NEWS'
+const init = {
+  selectedTab: 'NEWS',
+  nav: FootSectionNav.router.getStateForAction(FootSectionNav.router.getActionForPathAndParams('NEWS'))
 }
 
-export const footControl = (state = initState, action) => {
-  switch (action.type) {
-  case 'NEWS':
-    return {selectedTab: 'NEWS'};
-  case 'VIDEOS':
-    return {selectedTab: 'VIDEOS'};
-  case 'OTHERS':
-    return {selectedTab: 'OTHERS'};
-  }
-  return state;
+const appReducer  = (state = init, action) => {
+    switch(action.type){
+      case 'NEWS':
+        return {
+          selectedTab: 'NEWS',
+          nav: FootSectionNav.router.getStateForAction(FootSectionNav.router.getActionForPathAndParams('NEWS'))
+        };
+      case 'VIDEOS':
+        return {
+          selectedTab: 'VIDEOS',
+          nav: FootSectionNav.router.getStateForAction(FootSectionNav.router.getActionForPathAndParams('VIDEOS'))
+        };
+      case 'OTHERS':
+        return {
+          selectedTab: 'OTHERS',
+          nav: FootSectionNav.router.getStateForAction(FootSectionNav.router.getActionForPathAndParams('OTHERS'))
+        };
+    }
+    return state;
 }
 
-let store = createStore(footControl);
+let store = createStore(appReducer, init);
 
 export default store;
